Move tutorials and config
