+++ docs/tutorials/setup_iota-js_for_development.md
@@ -180,7 +180,7 @@ Now, you can go to the [Shimmer Testnet Faucet Website](https://faucet.testnet.s
 The 'Request' button will become clickable as soon as a valid Shimmer address is recognized.
 
 :::
-![Shimmer Testnet Faucet](/shimmer-testnet-faucet.png)
+![Shimmer Testnet Faucet](/img/tutorials/shimmer-testnet-faucet.png)
 
 ---
