+++ docs/tutorials/shimmerevm-hardhat.md
@@ -48,7 +48,7 @@ The sample project used here comes from the [Hardhat Quickstart guide](https://h
 
 You should see the following prompt:
 
-![Hardhat Setup](/shimmer-hardhat-setup.png 'Hardhat Setup')
+![Hardhat Setup](/img/tutorials/shimmer-hardhat-setup.png 'Hardhat Setup')
 
 Choose the JavaScript project and go through these steps to compile, test and deploy the sample contract.
 
@@ -128,12 +128,12 @@ Some important things to note are data types. `event` means that we can subscrib
 ```js
 module.exports = {
   solidity: '0.8.18',
-  defaultNetwork: 'shimmerevm-testnet',
+  defaultNetwork: 'shimmerevm',
   networks: {
-    'shimmerevm-testnet': {
-      url: 'https://json-rpc.evm.testnet.shimmer.network',
-      chainId: 1073,
-      accounts: [priv_key],
+    shimmerevm: {
+      url: 'https://json-rpc.evm.shimmer.network',
+      chainId: 148,
+      timeout: 60000,
     },
   },
 };
@@ -190,35 +190,37 @@ npx hardhat compile
 Run this command in root of the project directory:
 
 ```sh
-npx hardhat run scripts/deploy.js --network shimmerevm-testnet
+npx hardhat run scripts/deploy.js --network shimmerevm
 ```
 
-The contract will be deployed on ShimmerEVM Testnet, and you can check the deployment status here on the [explorer](https://explorer.evm.testnet.shimmer.network/).
+The contract will be deployed on ShimmerEVM, and you can check the deployment status here on the [explorer](https://explorer.evm.shimmer.network/).
 
 If you want to further verify your contract, add the following to your `hardhat.config.js`:
 
 ```js
 etherscan: {
-  apiKey: {
-    'shimmerevm-testnet': 'ABCDE12345ABCDE12345ABCDE123456789',
-  },
-  customChains: [
-    {
-      network: 'shimmerevm-testnet',
-      chainId: 1073,
-      urls: {
-        apiURL: 'https://explorer.evm.testnet.shimmer.network/api',
-        browserURL: 'https://explorer.evm.testnet.shimmer.network/',
+    apiKey:
+      {
+        shimmerevm: "xxx"
       },
-    },
-  ],
-}
+    customChains: [
+      {
+        apikey: "xxx",
+        network: "shimmerevm",
+        chainId: 148,
+        urls: {
+          apiURL: "https://explorer.evm.shimmer.network/api",
+          browserURL: "https://explorer.evm.shimmer.network/"
+        }
+      }
+    ]
+  },
 ```
 
 Then you can verify by running:
 
 ```sh
-npx hardhat verify --network shimmerevm-testnet <address> <unlock time>
+npx hardhat verify --network shimmerevm <address> <unlock time>
 ```
 
 :::tip Address and unlock time
@@ -228,7 +230,3 @@ You can easily copy the address and unlock time from the output of the [deploy s
 :::
 
 And Voila! You've successfully deployed your first Hardhat Smart Contract on ShimmerEVM and verified it on explorer.
-
-## Video Tutorial
-
-<iframe width="560" height="315" src="https://www.youtube.com/embed/rRscvKirroo" title="ShimmerEVM Tutorial: Deploying a Smart Contract with Hardhat" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
