+++ docs/tutorials/shimmerevm-setup.mdx
@@ -30,7 +30,7 @@ To add the ShimmerEVM network manually to your Metamask use following info:
 - Currency Symbol: `SMR`
 - Explorer URL: https://explorer.evm.shimmer.network
 
-![Metamask Adding New Network](/metamask-shimmerevm-mainnet.png 'Metamask Adding New Network')
+![Metamask Adding New Network](/img/tutorials/metamask-shimmerevm-mainnet.png 'Metamask Adding New Network')
 
 Hit `Save` and wait for Metamask to connect to the network.
