+++ docs/tutorials/shimmerevm-testnet-setup.mdx
@@ -27,7 +27,7 @@ To add the ShimmerEVM Testnet network manually to your Metamask use following in
 - Currency Symbol: `SMR`
 - Explorer URL: https://explorer.evm.testnet.shimmer.network
 
-![Metamask Adding New Network](/metamask-shimmerevm.png 'Metamask Adding New Network')
+![Metamask Adding New Network](/img/tutorials/metamask-shimmerevm.png 'Metamask Adding New Network')
 
 Alternatively, you can setup your own EVM Smart Contract Chain on top of the Shimmer network and provide a custom URL, which might look something like: `https://api.sc.testnet.shimmer.network/chain/rms1prr4r7az8e46qhagz5atugjm6x0xrg27d84677e3lurg0s6s76jr59dw4ls/evm/jsonrpc` as well as your own custom Chain ID.
 
@@ -38,10 +38,10 @@ At this point, our Metamask wallet is connected to the Shimmer EVM Testnet, but 
 ## Get EVM Account Funded
 
 Open [EVM Toolkit](https://evm-toolkit.evm.testnet.shimmer.network/) and enter your EVM address as shown below:
-![EVM Toolkit](/evm-toolkit.png 'EVM Toolkit')
+![EVM Toolkit](/img/tutorials/evm-toolkit.png 'EVM Toolkit')
 
 In case, you've setup a custom EVM chain, you'd need to add your custom network first. Please add your own chain address and select the relevant network as shown below:
-![EVM Toolkit Custom Network](/evm-toolkit-custom-network.png 'EVM Toolkit Custom Network')
+![EVM Toolkit Custom Network](/img/tutorials/evm-toolkit-custom-network.png 'EVM Toolkit Custom Network')
 
 Once you're done, click on `Send funds` and wait for a few minutes for the faucet to process. You should start seeing some funds on your Metamask wallet soon.
