+++ docusaurus.config.js
@@ -4,6 +4,7 @@ const { create_doc_plugin } = require('./src/utils/config');
 const contentConfigs = require('./contentPlugins');
 const articleRedirectsFile = require('./articleRedirects');
 const switcherConfig = require('./switcherConfig');
+const tutorials = require('./tutorials');
 const {
   buildPluginsConfig,
   maintainPluginsConfig,
@@ -210,6 +211,7 @@ module.exports = async () => {
         ],
       ],
       plugins: [
+        ...tutorials,
         ...contentPlugins,
         [
           '@docusaurus/plugin-content-docs',
