+++ src/utils/tags.ts
@@ -1,5 +1,5 @@
 //TODO tutorial.json should propably be a key-value dictionary
-import config from '../../tutorials/tags.json';
+import config from '../../tags.json';
 
 export interface Tag {
   label: string;
