+++ static/img/tutorials/connect-shimmerevm-and-deploy.png
