+++ static/img/tutorials/create-nft-with-wallet-lib.png
