+++ static/img/tutorials/proof-inclusion-of-a-block-architecture.png
