+++ static/img/tutorials/proof-inclusion-of-a-block.png
